# apps/fe/src/shared/api/openapi/api-1.yaml
openapi: 3.0.0
paths:
...
  /api/students:
    get:
      operationId: StudentsController_list
      parameters:
        - name: archived
          required: false
          in: query
          schema:
            type: string
            enum:
              - active
              - archived
              - all
      responses:
        "200":
          description: ""
      security: &a2
        - bearer: []
      tags: &a3
        - Students
    post:
      operationId: StudentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateStudentDto"
              example:
                displayName: Maksim Sokolov
                birthYear: 2012
                rating: 1460
                notes: Focuses on tactics and rook endgames.
      responses:
        "201":
          description: ""
      security: *a2
      tags: *a3
  /api/students/{studentId}:
    get:
      operationId: StudentsController_getOne
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
    patch:
      operationId: StudentsController_update
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateStudentDto"
              example:
                displayName: Maksim Sokolov
                birthYear: 2012
                rating: 1460
                notes: Focuses on tactics and rook endgames.
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
  /api/students/{studentId}/overview:
    get:
      operationId: StudentsController_getOverview
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/StudentOverviewResponse"
      security: *a2
      tags: *a3
  /api/students/{studentId}/analysis-profile:
    get:
      operationId: StudentsController_getAnalysisProfile
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/StudentAnalysisProfileResponse"
      security: *a2
      tags: *a3
  /api/students/{studentId}/performance-trend:
    get:
      operationId: StudentsController_getPerformanceTrend
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/StudentPerformanceTrendResponse"
      security: *a2
      tags: *a3
  /api/students/{studentId}/archive:
    post:
      operationId: StudentsController_setArchiveState
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SetStudentArchiveDto"
              example:
                archived: true
      responses:
        "200":
          description: ""
      security: *a2
      tags: *a3
...
  /api/analysis/{analysisId}/reports/generate:
    post:
      operationId: ReportsController_requestGeneration
      parameters:
        - name: analysisId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportGenerationRequestDto"
      responses:
        "201":
          description: ""
      security: &a12
        - bearer: []
      tags: &a13
        - Reports
  /api/reports:
    get:
      operationId: ReportsController_list
      parameters:
        - name: studentId
          required: true
          in: query
          schema:
            type: string
        - name: analysisId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a12
      tags: *a13
  /api/reports/{reportId}:
    get:
      operationId: ReportsController_getOne
      parameters:
        - name: reportId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a12
      tags: *a13
    patch:
      operationId: ReportsController_update
      parameters:
        - name: reportId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdateDto"
      responses:
        "200":
          description: ""
      security: *a12
      tags: *a13
    delete:
      operationId: ReportsController_remove
      parameters:
        - name: reportId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security: *a12
      tags: *a13
...
  /api/students/{studentId}/progress:
    get:
      operationId: ProgressController_getLatest
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ProgressReadResponse"
      security: &a16
        - bearer: []
      tags: &a17
        - Progress
  /api/students/{studentId}/progress/generate:
    post:
      operationId: ProgressController_requestGeneration
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: *a16
      tags: *a17
info:
  title: Chess Coach Copilot
  description: ""
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateStudentDto:
      type: object
      properties:
        displayName:
          type: string
          example: Maksim Sokolov
        birthYear:
          type: number
          example: 2012
        rating:
          type: number
          example: 1460
        notes:
          type: string
          example: Focuses on tactics and rook endgames.
      required:
        - displayName
    UpdateStudentDto:
      type: object
      properties:
        displayName:
          type: string
          example: Maksim Sokolov
        birthYear:
          type: number
          example: 2012
        rating:
          type: number
          example: 1460
        notes:
          type: string
          example: Focuses on tactics and rook endgames.
    SetStudentArchiveDto:
      type: object
      properties:
        archived:
          type: boolean
          example: true
      required:
        - archived
    StudentOverviewStudentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        coachAccountId:
          type: string
          format: uuid
        displayName:
          type: string
        birthYear:
          type: number
          nullable: true
        rating:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - coachAccountId
        - displayName
        - birthYear
        - rating
        - notes
        - archivedAt
        - createdAt
        - updatedAt
    ExternalAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        platform:
          type: string
          enum:
            - LICHESS
            - CHESS_COM
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - platform
        - username
        - createdAt
        - updatedAt
    StudentOverviewStatsResponse:
      type: object
      properties:
        gameCount:
          type: number
        analysisCount:
          type: number
        reportCount:
          type: number
        homeworkCount:
          type: number
      required:
        - gameCount
        - analysisCount
        - reportCount
        - homeworkCount
    LatestProgressReferenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisCount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisCount
        - createdAt
    RecentGameResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sourceLabel:
          type: string
          nullable: true
        studentColor:
          type: string
          enum:
            - WHITE
            - BLACK
        event:
          type: string
          nullable: true
        site:
          type: string
          nullable: true
        whitePlayerName:
          type: string
          nullable: true
        blackPlayerName:
          type: string
          nullable: true
        openingHeader:
          type: string
          nullable: true
        ecoCode:
          type: string
          nullable: true
        rawResult:
          type: string
          nullable: true
        derivedResult:
          type: string
          enum:
            - WIN
            - LOSS
            - DRAW
            - UNKNOWN
        plyCount:
          type: number
          nullable: true
        importedAt:
          type: string
          format: date-time
        latestAnalysisJobStatus:
          type: string
          enum:
            - PENDING
            - PARSING
            - EXTRACTING_ANNOTATIONS
            - CLASSIFICATION
            - GENERATING_OUTPUT
            - COMPLETED
            - FAILED
          nullable: true
        latestAnalysisJobId:
          type: string
          format: uuid
          nullable: true
        latestAnalysisId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - sourceLabel
        - studentColor
        - event
        - site
        - whitePlayerName
        - blackPlayerName
        - openingHeader
        - ecoCode
        - rawResult
        - derivedResult
        - plyCount
        - importedAt
        - latestAnalysisJobStatus
        - latestAnalysisJobId
        - latestAnalysisId
    RecentAnalysisResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        mainWeaknessTag:
          type: string
          enum:
            - MISSED_FORK
            - MISSED_PIN
            - MISSED_DOUBLE_ATTACK
            - MISSED_DISCOVERED_ATTACK
            - MISSED_MATE
            - ALLOWED_MATE
            - HANGING_PIECE
            - MISSED_CAPTURE
            - BAD_CAPTURE
            - MISSED_OPPONENT_THREAT
            - OPENED_LINE
            - UNKNOWN_TACTICAL_ERROR
            - KING_SAFETY
            - DELAYED_CASTLING
            - POOR_DEVELOPMENT
            - EARLY_QUEEN
            - BAD_TRADE
            - POOR_CONVERSION
            - OPENING_STRATEGY
            - PAWN_STRUCTURE
            - ENDGAME_TECHNIQUE
            - UNKNOWN_STRATEGIC_ERROR
            - TIME_MANAGEMENT
            - CALCULATION_DEPTH
            - TUNNEL_VISION
            - MATERIAL_GREED
            - LOW_BOARD_AWARENESS
            - INSUFFICIENT_ANNOTATION_DATA
            - REDUCED_CONFIDENCE
          nullable: true
        recommendedLessonTitle:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisJobId
        - gameId
        - mainWeaknessTag
        - recommendedLessonTitle
        - createdAt
    RecentReportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        title:
          type: string
        audience:
          type: string
          enum:
            - COACH
            - PARENT
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - title
        - audience
        - createdAt
    RecentHomeworkResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - title
        - createdAt
    StudentOverviewResponse:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/StudentOverviewStudentResponse"
        externalAccounts:
          type: array
          items:
            $ref: "#/components/schemas/ExternalAccountResponse"
        stats:
          $ref: "#/components/schemas/StudentOverviewStatsResponse"
        latestProgress:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/LatestProgressReferenceResponse"
        recentGames:
          type: array
          items:
            $ref: "#/components/schemas/RecentGameResponse"
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/RecentAnalysisResponse"
        recentReports:
          type: array
          items:
            $ref: "#/components/schemas/RecentReportResponse"
        recentHomework:
          type: array
          items:
            $ref: "#/components/schemas/RecentHomeworkResponse"
      required:
        - student
        - externalAccounts
        - stats
        - latestProgress
        - recentGames
        - recentAnalyses
        - recentReports
        - recentHomework
    WeaknessTagCountResponse:
      type: object
      properties:
        tag:
          type: string
          enum:
            - MISSED_FORK
            - MISSED_PIN
            - MISSED_DOUBLE_ATTACK
            - MISSED_DISCOVERED_ATTACK
            - MISSED_MATE
            - ALLOWED_MATE
            - HANGING_PIECE
            - MISSED_CAPTURE
            - BAD_CAPTURE
            - MISSED_OPPONENT_THREAT
            - OPENED_LINE
            - UNKNOWN_TACTICAL_ERROR
            - KING_SAFETY
            - DELAYED_CASTLING
            - POOR_DEVELOPMENT
            - EARLY_QUEEN
            - BAD_TRADE
            - POOR_CONVERSION
            - OPENING_STRATEGY
            - PAWN_STRUCTURE
            - ENDGAME_TECHNIQUE
            - UNKNOWN_STRATEGIC_ERROR
            - TIME_MANAGEMENT
            - CALCULATION_DEPTH
            - TUNNEL_VISION
            - MATERIAL_GREED
            - LOW_BOARD_AWARENESS
            - INSUFFICIENT_ANNOTATION_DATA
            - REDUCED_CONFIDENCE
        count:
          type: number
      required:
        - tag
        - count
    SeverityCountResponse:
      type: object
      properties:
        severity:
          type: string
          enum:
            - INACCURACY
            - MISTAKE
            - BLUNDER
            - MATE
            - UNKNOWN
        count:
          type: number
      required:
        - severity
        - count
    SampleMistakeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        severity:
          type: string
          enum:
            - INACCURACY
            - MISTAKE
            - BLUNDER
            - MATE
            - UNKNOWN
        category:
          type: string
        explanation:
          type: string
        suggestedFix:
          type: string
          nullable: true
      required:
        - id
        - analysisId
        - gameId
        - severity
        - category
        - explanation
        - suggestedFix
    StudentAnalysisProfileResponse:
      type: object
      properties:
        analysisCountUsed:
          type: number
        mainWeaknessTag:
          type: string
          enum:
            - MISSED_FORK
            - MISSED_PIN
            - MISSED_DOUBLE_ATTACK
            - MISSED_DISCOVERED_ATTACK
            - MISSED_MATE
            - ALLOWED_MATE
            - HANGING_PIECE
            - MISSED_CAPTURE
            - BAD_CAPTURE
            - MISSED_OPPONENT_THREAT
            - OPENED_LINE
            - UNKNOWN_TACTICAL_ERROR
            - KING_SAFETY
            - DELAYED_CASTLING
            - POOR_DEVELOPMENT
            - EARLY_QUEEN
            - BAD_TRADE
            - POOR_CONVERSION
            - OPENING_STRATEGY
            - PAWN_STRUCTURE
            - ENDGAME_TECHNIQUE
            - UNKNOWN_STRATEGIC_ERROR
            - TIME_MANAGEMENT
            - CALCULATION_DEPTH
            - TUNNEL_VISION
            - MATERIAL_GREED
            - LOW_BOARD_AWARENESS
            - INSUFFICIENT_ANNOTATION_DATA
            - REDUCED_CONFIDENCE
          nullable: true
        secondaryWeaknessTags:
          type: array
          items:
            type: string
            enum:
              - MISSED_FORK
              - MISSED_PIN
              - MISSED_DOUBLE_ATTACK
              - MISSED_DISCOVERED_ATTACK
              - MISSED_MATE
              - ALLOWED_MATE
              - HANGING_PIECE
              - MISSED_CAPTURE
              - BAD_CAPTURE
              - MISSED_OPPONENT_THREAT
              - OPENED_LINE
              - UNKNOWN_TACTICAL_ERROR
              - KING_SAFETY
              - DELAYED_CASTLING
              - POOR_DEVELOPMENT
              - EARLY_QUEEN
              - BAD_TRADE
              - POOR_CONVERSION
              - OPENING_STRATEGY
              - PAWN_STRUCTURE
              - ENDGAME_TECHNIQUE
              - UNKNOWN_STRATEGIC_ERROR
              - TIME_MANAGEMENT
              - CALCULATION_DEPTH
              - TUNNEL_VISION
              - MATERIAL_GREED
              - LOW_BOARD_AWARENESS
              - INSUFFICIENT_ANNOTATION_DATA
              - REDUCED_CONFIDENCE
        tagCounts:
          type: array
          items:
            $ref: "#/components/schemas/WeaknessTagCountResponse"
        severityCounts:
          type: array
          items:
            $ref: "#/components/schemas/SeverityCountResponse"
        sampleMistakes:
          type: array
          items:
            $ref: "#/components/schemas/SampleMistakeResponse"
        recommendedLessonTitle:
          type: string
          nullable: true
      required:
        - analysisCountUsed
        - mainWeaknessTag
        - secondaryWeaknessTags
        - tagCounts
        - severityCounts
        - sampleMistakes
        - recommendedLessonTitle
    PerformanceTrendPointResponse:
      type: object
      properties:
        date:
          type: string
          example: 2026-08-13
        value:
          type: number
      required:
        - date
        - value
    StudentPerformanceTrendResponse:
      type: object
      properties:
        direction:
          type: string
          enum:
            - IMPROVING
            - STABLE
            - DECLINING
            - UNKNOWN
        primaryMetric:
          type: string
        range:
          type: string
          enum:
            - 90D
        points:
          type: array
          items:
            $ref: "#/components/schemas/PerformanceTrendPointResponse"
      required:
        - direction
        - primaryMetric
        - range
        - points
...
    ReportGenerationRequestDto:
      type: object
      properties:
        audience:
          type: string
          enum:
            - COACH
            - PARENT
          example: COACH
      required:
        - audience
    ReportUpdateDto:
      type: object
      properties:
        title:
          type: string
          example: "Coach report: Italian Game review"
        content:
          type: object
          example:
            summary: Middlegame planning improved, but tactical checks were rushed.
            highlights:
              - Better opening recall
              - Missed forcing continuation on move 18
            lessonFocus:
              - Candidate move discipline
            nextSteps:
              - Review forcing-move checklist
...
    ProgressSnapshotResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        analysisCount:
          type: number
        summary:
          type: object
          additionalProperties: true
        promptVersion:
          type: string
        model:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - analysisCount
        - summary
        - promptVersion
        - model
        - createdAt
        - updatedAt
    ProgressReadResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - ready
            - not-enough-data
        requiredAnalysisCount:
          type: number
        availableAnalysisCount:
          type: number
        snapshot:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/ProgressSnapshotResponse"
      required:
        - status
        - requiredAnalysisCount
        - availableAnalysisCount
        - snapshot